// src/app/enums/actions.enum.ts
export enum Actions {

  CLEAR_ACTION_RESULT = 'CLEAR_ACTION_RESULT',

  CREATE_GROUP = 'CREATE_GROUP',
  CREATE_GROUP_COMPLETE = 'CREATE_GROUP_COMPLETE',
  UPDATE_GROUP = 'UPDATE_GROUP',
  UPDATE_GROUP_COMPLETE = 'UPDATE_GROUP_COMPLETE',
  DELETE_GROUP = 'DELETE_GROUP',
  DELETE_GROUP_COMPLETE = 'DELETE_GROUP_COMPLETE',
  FETCH_GROUP = 'FETCH_GROUP',
  FETCH_GROUP_COMPLETE = 'FETCH_GROUP_COMPLETE',
  FETCH_GROUPS = 'FETCH_GROUPS',
  FETCH_GROUPS_COMPLETE = 'FETCH_GROUPS_COMPLETE',
  FETCH_GROUP_MEMBERS_COMPLETE = 'FETCH_GROUP_MEMBERS_COMPLETE',
  FETCH_GROUP_MEMBERS = 'FETCH_GROUP_MEMBERS',
  DELETE_GROUP_MEMBERS_COMPLETE = 'DELETE_GROUP_MEMBERS_COMPLETE',
  DELETE_GROUP_MEMBERS = 'DELETE_GROUP_MEMBERS',
  DELETE_GROUP_MEMBER_COMPLETE = 'DELETE_GROUP_MEMBER_COMPLETE',
  DELETE_GROUP_MEMBER = 'DELETE_GROUP_MEMBER',
  ADD_GROUP_MEMBER = 'ADD_GROUP_MEMBER',
  ADD_GROUP_MEMBER_COMPLETE = 'ADD_GROUP_MEMBER_COMPLETE',
  ADD_GROUP_MEMBERS = 'ADD_GROUP_MEMBERS',
  ADD_GROUP_MEMBERS_COMPLETE = 'ADD_GROUP_MEMBERS_COMPLETE',

  UPDATE_USER_COMPLETE = 'UPDATE_USER_COMPLETE',
  UPDATE_USER = 'UPDATE_USER',
  CREATE_USER_COMPLETE = 'CREATE_USER_COMPLETE',
  CREATE_USER = 'CREATE_USER',
  DELETE_USER = 'DELETE_USER',
  DELETE_USER_COMPLETE = 'DELETE_USER_COMPLETE',
  FETCH_USERS_COMPLETE = 'FETCH_USERS_COMPLETE',
  FETCH_USERS = 'FETCH_USERS',
  FETCH_USER_COMPLETE = 'FETCH_USER_COMPLETE',
  FETCH_USER = 'FETCH_USER',

  STUDIO_INIT = 'STUDIO_INIT',

  LOGIN = 'LOGIN',
  LOGGED_IN = 'LOGGED_IN',
  LOGOUT = 'LOGOUT',
  LOGGED_OUT = 'LOGGED_OUT',
  SESSION_TIMEOUT = 'SESSION_TIMEOUT',
  RECOVER_PASSWORD = 'RECOVER_PASSWORD',

  UPDATE_SETTING = 'UPDATE_SETTING',
  UPDATE_SETTINGS = 'UPDATE_SETTINGS',
  SHOW_SIDEBAR = 'SHOW_SIDEBAR',
  HIDE_SIDEBAR = 'HIDE_SIDEBAR',
  TOGGLE_SIDEBAR = 'TOGGLE_SIDEBAR',
  TOGGLE_SIDEBAR_FOLD = 'TOGGLE_SIDEBAR_FOLD',

  // selectedItems
  SELECT_ITEM = 'SELECT_ITEM',
  DESELECT_ITEM = 'DESELECT_ITEM',
  DESELECT_ITEMS = 'DESELECT_ITEMS',
  SELECT_ITEMS = 'SELECT_ITEMS',

  // expandedPanels
  EXPAND_PANEL = 'EXPAND_PANEL',
  COLLAPSE_PANEL = 'COLLAPSE_PANEL',
  EXPAND_PANELS = 'EXPAND_PANELS',
  COLLAPSE_PANELS = 'COLLAPSE_PANELS',

  // expandedPaths
  EXPAND_PATH = 'EXPAND_PATH',
  COLLAPSE_PATH = 'COLLAPSE_PATH',
  EXPAND_PATHS = 'EXPAND_PATHS',
  COLLAPSE_PATHS = 'COLLAPSE_PATHS',

  // previewTabs
  OPEN_TAB = 'OPEN_TAB',
  OPEN_TABS = 'OPEN_TABS',
  CLOSE_TAB = 'CLOSE_TAB',
  SELECT_TAB = 'SELECT_TAB',
  GUEST_CHECK_IN = 'GUEST_CHECK_IN',
  NAVIGATE_ON_ACTIVE = 'NAVIGATE_ON_ACTIVE',
  OPEN_TAB_BACKGROUND = 'OPEN_TAB_BACKGROUND',
  OPEN_TABS_BACKGROUND = 'OPEN_TABS_BACKGROUND',
  TAB_NAVIGATE_BACK = 'TAB_NAVIGATE_BACK',
  TAB_NAVIGATE_FORWARD = 'TAB_NAVIGATE_FORWARD',

  // projects
  FETCH_PROJECTS = 'FETCH_PROJECTS',
  FETCH_PROJECT = 'FETCH_PROJECT',
  PROJECTS_FETCHED = 'PROJECTS_FETCHED',
  PROJECT_FETCHED = 'PROJECT_FETCHED',
  PROJECTS_FETCH_ERROR = 'PROJECTS_FETCH_ERROR',
  PROJECT_FETCH_ERROR = 'PROJECT_FETCH_ERROR',
  CREATE_PROJECT = 'CREATE_PROJECT',
  UPDATE_PROJECT = 'UPDATE_PROJECT',
  DELETE_PROJECT = 'DELETE_PROJECT',
  PROJECT_CREATED = 'PROJECT_CREATED',
  PROJECT_UPDATED = 'PROJECT_UPDATED',
  PROJECT_DELETED = 'PROJECT_DELETED',
  EDIT_PROJECT = 'EDIT_PROJECT',

  SELECT_PROJECT = 'SELECT_PROJECT',
  DESELECT_PROJECT = 'DESELECT_PROJECT',

  // assets
  FETCH_ASSETS = 'FETCH_ASSETS',
  FETCH_SOME_ASSETS = 'FETCH_SOME_ASSETS',
  FETCH_ASSET = 'FETCH_ASSET',
  ASSETS_FETCHED = 'ASSETS_FETCHED',
  SOME_ASSETS_FETCHED = 'SOME_ASSETS_FETCHED',
  ASSET_FETCHED = 'ASSET_FETCHED',
  ASSETS_FETCH_ERROR = 'ASSETS_FETCH_ERROR',
  ASSET_FETCH_ERROR = 'ASSET_FETCH_ERROR',
  CREATE_ASSET = 'CREATE_ASSET',
  UPDATE_ASSET = 'UPDATE_ASSET',
  DELETE_ASSET = 'DELETE_ASSET',
  ASSET_CREATED = 'ASSET_CREATED',
  ASSET_UPDATED = 'ASSET_UPDATED',
  ASSET_DELETED = 'ASSET_DELETED',
  EDIT_ASSETS = 'EDIT_ASSETS',
  EDIT_ASSET = 'EDIT_ASSET',
  FETCH_ASSET_FOR_EDIT = 'FETCH_ASSET_FOR_EDIT',
  ASSET_FETCHED_FOR_EDIT = 'ASSET_FETCHED_FOR_EDIT',
  UPDATE_EDIT_SESSION = 'UPDATE_EDIT_SESSION',
  CLOSE_EDIT_SESSION = 'CLOSE_EDIT_SESSION',
  EDIT_SESSION_CLOSED = 'EDIT_SESSION_CLOSED',
  CHANGE_ACTIVE_EDIT_SESSION = 'CHANGE_ACTIVE_EDIT_SESSION',
  SESSION_CHANGES_PERSISTED = 'SESSION_CHANGES_PERSISTED',
  PERSIST_SESSION_CHANGES = 'PERSIST_SESSION_CHANGES',

  EXPLORER_SELECT_PROJECT = 'EXPLORER_SELECT_PROJECT',
  EXPLORER_SELECT_PATH = 'EXPLORER_SELECT_PATH',
  EXPLORER_SELECT_ASSET = 'EXPLORER_SELECT_ASSET',

}